pyproject.toml
[project]
name = "institution-design"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pyyaml>=6.0",
    "numba>=0.60.0",
]
